requirements.txt
absl-py==1.4.0
astunparse==1.6.3
blinker==1.6.2
boto3==1.26.134
botocore==1.29.134
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.1
Flask-Cors==3.0.10
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
protobuf
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydub==0.25.1
openai==0.27.2
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
termcolor==2.3.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.0
wrapt==1.14.1
opencv-python-headless==4.9.0.80
mediapipe==0.9.1.0

